Be more defensive about path structures
As Will pointed out there are a lot of potential inputs that we're being passed and from that we need to figure out the cookbook root. Outside of walking the directory structure looking for metadata files and somehow excluding test cookbooks this is pretty difficult. I believe I have a working solution that is a lot less brittle for this. I added specs for the 2 extra scenarios we need to account for.

Signed-off-by: Tim Smith <tsmith@chef.io>