Add new --without-search flag to tidy server report
This causes the command to load the nodes list using the /nodes
API instead of the partial search API.  This will result in much longer
run times, and should only be used if the organization has more nodes
than can be retrieved in a search (~10000)

This is a short-term workaround; the better approach will probably be to
switch over to use of 'point in time' in the open/elasticsearch search API,
which will allow for safe retention of large resultsets for query.

If we keep this change, we could improve it further by automatically detecting
which mode to use (nodes v search) based on node count.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>